+++ assets/javascripts/kitten/karl/layout/separator.js
@@ -1,0 +1,15 @@
+import React from 'react'
+import { Marger } from 'kitten/components/layout/marger'
+import { Separator } from 'kitten/components/layout/separator'
+
+export const KarlSeparator = props => (
+  <Marger top="5">
+    <Separator />
+  </Marger>
+)
+
+export const KarlSeparatorDarker = props => (
+  <Marger top="5">
+    <Separator darker={ true } />
+  </Marger>
+)