+++ spec/dummy/client/entries/app-kitten.js
@@ -178,6 +178,8 @@ import { KarlMenuHeader } from 'kitten/karl/headers/karl-header'
 import { KarlMenu,
          KarlMenuContent } from 'kitten/karl/menus/karl-menu'
 import { KarlBrowsing } from 'app/browsing/browsing'
+import { KarlSeparator,
+         KarlSeparatorDarker } from 'kitten/karl/layout/separator'
 
 ReactOnRails.register({
   // Box
@@ -352,4 +354,6 @@ ReactOnRails.register({
   KarlMenu,
   KarlMenuContent,
   KarlBrowsing,
+  KarlSeparator,
+  KarlSeparatorDarker,
 })